rename to AutoSaveStatus
