--- src/AutoSave/index.ts
@@ -1,2 +1,0 @@
-export { AutoSaveContext, AutoSaveProvider } from "./AutoSaveProvider";
-export { useAutoSave } from "./useAutoSave";
