+++ src/index.ts
@@ -5,4 +5,4 @@ export { FieldState } from "src/fields/valueField";
 export { required, Rule } from "src/rules";
 export { useFormState } from "src/useFormState";
 export { useFormStates } from "src/useFormStates";
-export * from "./AutoSave";
+export * from "./AutoSaveStatus";
