+++ src/useFormState.ts
@@ -2,7 +2,7 @@ import { isPlainObject } from "is-plain-object";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { ObjectConfig } from "src/config";
 import { createObjectState, ObjectState } from "src/fields/objectField";
-import { useAutoSave } from "./AutoSave";
+import { useAutoSaveStatus } from "./AutoSaveStatus";
 import { initValue, isInput, isQuery } from "./utils";
 
 // A structural match for useQuery
@@ -83,7 +83,7 @@ let pendingAutoSave = false;
  */
 export function useFormState<T, I>(opts: UseFormStateOpts<T, I>): ObjectState<T> {
   const { config, init, addRules, readOnly = false, loading, autoSave } = opts;
-  const autoSaveContext = useAutoSave();
+  const autoSaveStatusContext = useAutoSaveStatus();
 
   // Use a ref so our memo'ized `onBlur` always see the latest value
   const autoSaveRef = useRef<((state: ObjectState<T>) => void) | undefined>(autoSave);
@@ -115,11 +115,11 @@ export function useFormState<T, I>(opts: UseFormStateOpts<T, I>): ObjectState<T>
               // user's autoSave function itself wants to call a .set.
               const promise = autoSaveRef.current!(form);
               isAutoSaving = "in-flight";
-              autoSaveContext.triggerAutoSave();
+              autoSaveStatusContext.triggerAutoSave();
               await promise;
             } finally {
               isAutoSaving = false;
-              autoSaveContext.resolveAutoSave();
+              autoSaveStatusContext.resolveAutoSave();
               if (pendingAutoSave) {
                 pendingAutoSave = false;
                 // Push out the follow-up by 1 tick to allow refreshes to happen to potentially
